requirements.txt
annotated-types==0.7.0
APScheduler==3.11.0
blinker==1.8.2
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.10
dateparser==1.2.0
DAWG-Python==0.7.2
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
fsspec==2024.10.0
fuzzywuzzy==0.18.0
huggingface-hub==0.26.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
Levenshtein==0.26.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.11
networkx==3.4.2
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
preshed==3.0.9
psycopg2==2.9.10
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
pymorphy3==2.0.2
pymorphy3-dicts-ru==2.4.417150.4580142
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ru_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.8.0/ru_core_news_sm-3.8.0-py3-none-any.whl#sha256=69978d47b43e2c4f329bebdb155e8e9d3861bba1a58ba25551419dae7d7e07fc
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.13.1
thinc==8.3.2
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.1
wrapt==1.17.0
